REname some traits
