+++ src/Patpass/PatpassComercio/Inscription.php
@@ -14,7 +14,7 @@ use Transbank\Patpass\PatpassComercio\Exceptions\InscriptionStatusException;
 use Transbank\Patpass\PatpassComercio\Responses\InscriptionStartResponse;
 use Transbank\Patpass\PatpassComercio\Responses\InscriptionStatusResponse;
 use Transbank\Webpay\Exceptions\WebpayRequestException;
-use Transbank\Webpay\InteractsWithWebpayApi;
+use Transbank\Utils\InteractsWithWebpayApi;
 use Transbank\Webpay\Options;
 
 class Inscription
