+++ src/Utils/ConfiguresEnvironment.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace Transbank\Webpay;
+namespace Transbank\Utils;
 
 use Transbank\Utils\HttpClient;
+use Transbank\Webpay\Options;
 
 trait ConfiguresEnvironment
 {
